Able to calculate molecular mass, and determine limiting reagent

+++ src/main/java/com/fezda/chemistry/Element.java
@@ -1,14 +1,14 @@
 package com.fezda.chemistry;
 
 public class Element {
-	private String name;
-	private String symbol;
-	private int number;
-	private float weight;
-	private int group;
-	private int period;
+	private String name;		// Full name
+	private String symbol;		// Two-letter representation
+	private int number;			// Atomic number
+	private double weight;		// Atomic weight (g/mol)
+	private int group;			// Column number in periodic table
+	private int period;			// Row num in periodic table
 	
-	public Element (String name, String symbol, int number, float weight, int group, int period) {
+	public Element (String name, String symbol, int number, double weight, int group, int period) {
 		this.name = name;
 		this.symbol = symbol;
 		this.number = number;
@@ -29,7 +29,7 @@ public class Element {
 		return number;
 	}
 
-	public float getWeight() {
+	public double getWeight() {
 		return weight;
 	}
 	
@@ -37,7 +37,7 @@ public class Element {
 		return group;
 	}
 
-	public float getPeriod() {
+	public int getPeriod() {
 		return period;
 	}
